feat: 반응형 작업 세팅, 페이지 컴포넌트 분리
