+++ src/App.jsx
@@ -1,15 +1,15 @@
 import { Routes, Route } from 'react-router-dom';
-import Home from './components/pages/HomePage/HomePage';
 import Layout from './components/atoms/Layout/Layout';
-import Game from './components/pages/GamePage/GamePage';
-import Chat from './components/pages/GamePage/ChatPage';
+import Chat from './components/pages/ChatPage/ChatPage';
+import Login from './components/pages/LoginPage/LoginPage';
+import Draw from './components/pages/DrawPage/DrawPage';
 
 function App() {
   return (
     <Routes>
       <Route path="/" element={<Layout />}>
-        <Route path="/" element={<Home />} />
-        <Route path="/game" element={<Game />} />
+        <Route path="/" element={<Login />} />
+        <Route path="/game" element={<Draw />} />
         <Route path="/chat" element={<Chat />} />
       </Route>
     </Routes>
