+++ src/assets/css/common.css
@@ -26,10 +26,9 @@
 
 /* Reset */
 body *{box-sizing: border-box; outline: none;}
-body.fixed{overflow-y: hidden;}
 html,body,h1,h2,h3,h4,h5,h6,div,p,blockquote,pre,code,address,ul,ol,li,menu,nav,section,article,aside,dl,dt,dd,table,thead,tbody,tfoot,label,caption,th,td,form,fieldset,legend,hr,input,button,textarea,object,figure,figcaption{margin:0;padding:0;}
-body{-webkit-text-size-adjust:none;word-wrap:break-word;word-break:break-all;background:#fff;}
-body,input,select,textarea,button{border:none; font-size:14px; font-family: 'Noto Sans KR', sans-serif; color:#333333;}
+body{-webkit-text-size-adjust:100%;word-wrap:break-word;word-break:break-all;background:#fff;}
+body,input,select,textarea,button{border:none; font-size:1.4rem; font-family: 'Noto Sans KR', sans-serif; color:#333333;}
 ul,ol,li{list-style:none;}
 table{border-spacing:0;}
 img,fieldset{border:0;}
@@ -40,9 +39,8 @@ label,img,input,select,textarea,button{vertical-align:middle;}
 hr{display:none;}
 main,header,section,nav,footer,aside,article,figure{display:block;}
 a{color:#333333;text-decoration:none;}
-button{cursor: pointer; color: #333333; background-color: transparent; outline: none; font-size: 16px;}
+button{cursor: pointer; color: #333333; background-color: transparent; outline: none; font-size: 1.6rem;}
 caption{position: relative;}
-#wrap{min-width:1400px;}
 
 /* Form */
 input[type=password],
@@ -54,18 +52,18 @@ input[type=email],
 input[type=file],
 input[type=url],
 input[type=number],
-input[type=date],textarea{font-size:14px; color:#333333; border:1px solid #e0e0e0; background-color:#fff; transition: all 0.5s; box-sizing: border-box; vertical-align:middle; border-radius: 4px; -webkit-appearance: none; -moz-appearance: none; appearance: none;}
+input[type=date],textarea{font-size:1.4rem; color:#333333; border:1px solid #e0e0e0; background-color:#fff; transition: all 0.5s; box-sizing: border-box; vertical-align:middle; border-radius: 4px; -webkit-appearance: none; -moz-appearance: none; appearance: none;}
 pre{white-space: -moz-pre-wrap; white-space: -pre-wrap; white-space: -o-pre-wrap; white-space: pre-wrap; word-wrap: break-word; font-family: 'Noto Sans KR', sans-serif;}
 input[type=password]::-moz-reveal,
 input[type=password]::-ms-reveal{display: none; width: 0; height: 0;}
 input[type=email],
 input[type=number],
 input[type=password],
-input[type=text] {padding: 0 16px; height: 40px; line-height: 40px; border-radius: 4px;}
-input::-webkit-input-placeholder{color:#c4c4c4; font-size:14px; line-height:100%;}
+input[type=text] {padding: 0 1.6rem; height: 4rem; line-height: 4rem; border-radius: 4px;}
+input::-webkit-input-placeholder{color:#c4c4c4; font-size:1.4rem; line-height:100%;}
 input:not([type=range]):focus,
 textarea:focus {border: 1px solid #679DFF !important;}
-textarea{padding: 10px 16px; border:1px solid #e1e1e1;}
+textarea{padding: 1rem 1.6rem; border:1px solid #e1e1e1;}
 textarea::-webkit-input-placeholder{color:#c4c4c4;}
 input[type="number"]::-webkit-outer-spin-button,
 input[type="number"]::-webkit-inner-spin-button{-webkit-appearance: none; margin: 0;}
@@ -79,15 +77,15 @@ input[type=tel][readonly],
 input[type=text][readonly],
 input[type=password][readonly],
 input[type=email][readonly],
-input[type=search][readonly]{-webkit-appearance:none; font-size:14px; padding: 0 16px; background: #f8f8f8; border: 1px solid #e0e0e0; color: #333333;}
+input[type=search][readonly]{-webkit-appearance:none; font-size:1.4rem; padding: 0 1.6rem; background: #f8f8f8; border: 1px solid #e0e0e0; color: #333333;}
 input[type=tel][disabled],
 input[type=text][disabled],
 input[type=password][disabled],
 input[type=search][disabled],
-input[type=email][disabled]{-webkit-appearance:none; font-size:14px; padding: 0 16px; background: #f8f8f8; border: 1px solid #e0e0e0; color: #c4c4c4;}
+input[type=email][disabled]{-webkit-appearance:none; font-size:1.4rem; padding: 0 1.6rem; background: #f8f8f8; border: 1px solid #e0e0e0; color: #c4c4c4;}
 select:disabled{color: #c4c4c4; background-color: #f8f8f8; border: 1px solid #e0e0e0; cursor: default;}
 textarea[readonly],
-textarea[disabled]{padding:12px;color:#e1e1e1; font-weight:normal; line-height:100%; height:40px; background:#f8f8f9;border:1px solid #e1e1e1;}
+textarea[disabled]{padding:1.2rem;color:#e1e1e1; font-weight:normal; line-height:100%; height:4rem; background:#f8f8f9;border:1px solid #e1e1e1;}
 .clear {clear:both;}
 .clear:after {content:""; display:block; clear:both;}
 .left{float: left;}
@@ -110,72 +108,35 @@ input:-webkit-autofill:active {
 .scroll::-webkit-scrollbar-thumb{background-color: #e0e0e0; border-radius: 100px;}
 .scroll::-webkit-scrollbar-track{background-color: #e5e5e5; border-radius: 100px; background-clip: padding-box; border: 3px solid transparent;}
 
-/* textarea 가로 고정 */
-textarea.re_ver{resize : vertical;}
-textarea.re_none{resize: none;}
+/* REM rules */
+html{font-size:3px !important;}
+@media screen and (min-width:216px){html{font-size:3.5px !important;}}
+@media screen and (min-width:229px){html{font-size:3.8px !important;}}
+@media screen and (min-width:250px){html{font-size:4.2px !important;}}
+@media screen and (min-width:252px){html{font-size:4.5px !important;}}
+@media screen and (min-width:288px){html{font-size:4.8px !important;}}
+@media screen and (min-width:300px){html{font-size:5px !important;}}
+/* iphone 5 */
+@media screen and (min-width:320px){html{font-size:6px !important;}}
+@media screen and (min-width:360px){html{font-size:7px !important;}}
+@media screen and (min-width:396px){html{font-size:8px !important;}}
+@media screen and (min-width:432px){html{font-size:8.4px !important;}}
+@media screen and (min-width:460px){html{font-size:8.8px !important;}}
+@media screen and (min-width:504px){html{font-size:9.2px !important;}}
+@media screen and (min-width:540px){html{font-size:9.5px !important;}}
+@media screen and (min-width:576px){html{font-size:9.8px !important;}}
+@media screen and (min-width:600px){html{font-size:10px !important;}}
+@media screen and (min-width:820px){html{font-size:10px !important;}}
+body {font-size:1rem !important;}
 
-/* text 말줄임표 */
-.ellipsis{white-space:nowrap;text-overflow:ellipsis;overflow:hidden}
-
-/* vertical-align */
-.ver_top{vertical-align: top;}
-.ver_mid{vertical-align: middle;}
-.ver_btm{vertical-align: bottom;}
-
-/* text-align */
-.txt_left{text-align: left;}
-.txt_center{text-align: center;}
-.txt_right{text-align: right;}
-
-/* display */
-.dis_block{display: block;}
-.dis_inblock{display: inline-block;}
-.dis_none{display: none;}
-
-/* 가로사이즈 */
-.w_64{width: 64px;}
-.w_100{width: 100px;}
-.w_120{width: 120px;}
-.w_156{width: 156px;}
-.w_200{width: 200px;}
-.w_276{width: 276px;}
-.w_400{width: 400px;}
-.w_full{width: 100%;}
-.w_minus210{width:calc(100% - 210px);}
-
-/* margin, padding */
-.marleft_4{margin-left: 4px;}
-.marleft_6{margin-left: 6px;}
-.marleft_10{margin-left: 10px;}
-.marleft_12{margin-left: 12px;}
-.marleft_40{margin-left: 40px;}
-
-.martop_4{margin-top: 4px;}
-.martop_8{margin-top: 8px;}
-.martop_10{margin-top: 10px;}
-.martop_20{margin-top: 20px;}
-.martop_24{margin-top: 24px;}
-.martop_40{margin-top: 40px;}
-
-.marright_10{margin-right: 10px;}
-
-/* 폰트컬러 */
-.col_blue{color:#4D47C3;}
-.col_grey{color:#666666;}
-
-/* background-color */
-.bg_grey{background-color: #f8f8f9;}
-
-/* 폰트 사이즈 */
-.fs_12{font-size: 12px;}
-.fs_14{font-size: 14px;}
-.fs_16{font-size: 16px;}
-.fs_18{font-size: 18px;}
-.fs_20{font-size: 20px;}
-
-/* font weight */
-.fw_normal{font-weight: 400;}
-.fw_bold{font-weight: 700;}
-.fw_ebold{font-weight: 900;}
-
-/* button */
+main {
+    min-height: 440px;
+    background: #f4f5f7 url('../images/bg@2x.png') no-repeat center top / 100%;
+    color: #212121;
+    font-family: 'Open Sans', sans-serif;
+}
+@media screen and (max-width:900px) {
+    main{
+        background: #f4f5f7 url('../images/bg@2x.png') no-repeat left top / cover
+    }
+}
