+++ src/components/atoms/Layout/Layout.jsx
@@ -1,17 +1,15 @@
 import { Outlet } from 'react-router-dom';
 
 const Layout = () => {
-    return (
-        <>
-            {/* <Header /> */}
-            <div id="wrap">
-                <main>
-                    <Outlet />
-                </main>
-                {/* <Footer /> */}
-            </div>
-        </>
-    );
+  return (
+    <>
+      <div id="wrap">
+        <main>
+          <Outlet />
+        </main>
+      </div>
+    </>
+  );
 };
 
 export default Layout;
