+++ src/components/pages/ChatPage/ChatPage.scss
@@ -1,33 +1,33 @@
 .chat_window {
-  width: 340px;
-  height: 420px;
-  border-radius: 24px;
+  width: 34rem;
+  height: 42rem;
+  border-radius: 2.4rem;
   overflow: hidden;
   height: 100%;
-  border: 1px solid #d8d8d8;
+  border: 0.1rem solid #d8d8d8;
   p {
     margin: 0;
-    font-size: 13px;
+    font-size: 1.3rem;
   }
   .chat_header {
-    height: 54px;
+    height: 5.4rem;
     position: relative;
-    border-bottom: 1px solid rgba(0, 15, 56, 0.1);
+    border-bottom: 0.1rem solid rgba(0, 15, 56, 0.1);
 
     p {
       position: relative;
-      padding-left: 54px;
+      padding-left: 5.4rem;
       display: block;
       font-weight: 700;
-      line-height: 54px;
+      line-height: 5.4rem;
       &::before {
         content: '';
         position: absolute;
         top: 50%;
         transform: translateY(-50%);
-        left: 16px;
-        width: 28px;
-        height: 28px;
+        left: 1.6rem;
+        width: 2.8rem;
+        height: 2.8rem;
         background: url('../../../assets/images/icons/profile_default.png')
           no-repeat center / contain;
       }
@@ -35,7 +35,7 @@
   }
 
   .chat_body {
-    height: calc(100% - 104px);
+    height: calc(100% - 10.4rem);
     background: #fff;
     position: relative;
 
@@ -47,49 +47,49 @@
     }
 
     .message_container::-webkit-scrollbar {
-      width: 4px;
-      height: 4px;
+      width: 0.4rem;
+      height: 0.4rem;
     }
     .message_container::-webkit-scrollbar-track {
       background-color: #525275;
-      border-radius: 100px;
+      border-radius: 10rem;
       background-clip: padding-box;
-      border: 3px solid transparent;
+      border: 0.3rem solid transparent;
     }
     .message_container::-webkit-scrollbar-thumb {
       background: #f2f2f6;
-      border-radius: 100px;
+      border-radius: 10rem;
     }
 
     .message {
       height: auto;
-      padding: 10px;
+      padding: 1rem;
       display: flex;
 
       .message_content {
         width: auto;
         height: auto;
-        min-height: 40px;
-        max-width: 120px;
+        min-height: 4rem;
+        max-width: 12rem;
         background-color: rgba($color: #679dff, $alpha: 1);
-        border-radius: 5px;
+        border-radius: 0.5rem;
         color: white;
         display: flex;
         align-items: center;
-        margin-right: 5px;
-        margin-left: 5px;
-        padding: 8px;
+        margin-right: 0.5rem;
+        margin-left: 0.5rem;
+        padding: 0.8rem;
         overflow-wrap: break-word;
         word-break: break-word;
       }
 
       .message_meta {
-        margin-top: 4px;
+        margin-top: 0.4rem;
         display: flex;
-        font-size: 12px;
+        font-size: 1.2rem;
 
         #author {
-          margin-right: 6px;
+          margin-right: 0.6rem;
           font-weight: bold;
         }
       }
@@ -97,24 +97,24 @@
   }
 
   .chat_footer {
-    margin-bottom: 10px;
-    padding: 0 10px;
+    margin-bottom: 1rem;
+    padding: 0 1rem;
     border-top: none;
     display: flex;
     justify-content: space-around;
 
     input {
-      width: 260px;
-      border-radius: 12px;
+      width: 26rem;
+      border-radius: 1.2rem;
     }
 
     button {
-      width: 40px;
-      height: 40px;
+      width: 4rem;
+      height: 4rem;
       background: #242e47;
-      font-size: 16px;
+      font-size: 1.6rem;
       color: #fff;
-      border-radius: 12px;
+      border-radius: 1.2rem;
     }
     button:hover {
       background: #242e47e5;
@@ -127,15 +127,15 @@
 
   .message_content {
     justify-content: flex-end;
-    box-shadow: 4px 4px 12px rgb(0, 15, 56, 0.2);
+    box-shadow: 0.4rem 0.4rem 1.2rem rgb(0, 15, 56, 0.2);
   }
 
   .message_meta {
     justify-content: flex-end;
-    margin-left: 5px;
+    margin-left: 0.5rem;
 
     #time {
-      margin-right: 6px;
+      margin-right: 0.6rem;
     }
   }
 }
@@ -147,17 +147,17 @@
     justify-content: flex-start;
     background-color: #fff;
     color: #333;
-    box-shadow: 4px 4px 12px rgb(0, 15, 56, 0.2);
+    box-shadow: 0.4rem 0.4rem 1.2rem rgb(0, 15, 56, 0.2);
     background-color: #e4e4e6;
   }
 
   .message_meta {
     justify-content: flex-end;
     flex-direction: row-reverse;
-    margin-right: 5px;
+    margin-right: 0.5rem;
 
     #author {
-      margin-left: 6px;
+      margin-left: 0.6rem;
       font-weight: bold;
     }
   }
