+++ src/components/pages/LoginPage/LoginPage.scss
@@ -1,44 +1,31 @@
 .home {
   height: 100vh;
-  min-height: 440px;
-  background: #f4f5f7 url('../../../assets/images/bg@2x.png') no-repeat center
-    top / 100%;
-  color: #212121;
-  font-family: 'Open Sans', sans-serif;
-  // display: grid;
-  // place-items: center;
-
   display: flex;
   align-items: center;
   justify-content: center;
 }
-
 .join_chat_container {
-  // position: absolute;
-  // top: 50%;
-  // left: 50%;
-  // transform: translate(-50%, -50%);
   position: relative;
   background-color: #fff;
-  width: 400px;
-  height: 400px;
-  padding: 0 60px;
+  width: 40rem;
+  height: 40rem;
+  padding: 0 6rem;
   display: flex;
   justify-content: center;
   align-items: center;
   flex-direction: column;
   text-align: center;
-  border-radius: 24px;
+  border-radius: 2.4rem;
 
   &::before {
     content: '';
     position: absolute;
     left: 0;
-    top: -100px;
-    width: 100px;
-    height: 100px;
+    top: -10rem;
+    width: 10rem;
+    height: 10rem;
     background: url('../../../assets/images/img_participants.png') no-repeat
-      center;
+      center / contain;
   }
 
   h3 {
@@ -47,23 +34,22 @@
   }
   input[type='text'] {
     width: 100%;
-    height: 50px;
-    margin: 10px;
-    // border: 2px solid #43a047;
-    border-radius: 5px;
+    height: 5rem;
+    margin: 1rem;
+    border-radius: 0.5rem;
   }
   button {
     width: 100%;
-    height: 50px;
-    margin: 10px;
+    height: 5rem;
+    margin: 1rem;
     border: none;
-    border-radius: 5px;
-    padding: 5px;
-    font-size: 16px;
+    border-radius: 0.5rem;
+    padding: 0.5rem;
+    font-size: 1.6rem;
     background: #242e47;
     color: #fff;
     cursor: pointer;
-    border-radius: 28px;
+    border-radius: 2.8rem;
   }
   button:hover {
     background: #242e47e5;
@@ -75,58 +61,58 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  width: 1200px;
+  width: 120rem;
 
   strong {
     display: inline-block;
-    padding: 0 50px;
-    height: 68px;
-    line-height: 68px;
-    margin-bottom: 24px;
-    font-size: 32px;
-    border-radius: 32px;
+    padding: 0 5rem;
+    height: 6.8rem;
+    line-height: 6.8rem;
+    margin-bottom: 2.4rem;
+    font-size: 3.2rem;
+    border-radius: 3.2rem;
     color: #fff;
     background-color: #3b80ff;
   }
 
   .participants {
-    padding: 16px 0;
-    margin-bottom: 24px;
+    padding: 1.6rem 0;
+    margin-bottom: 2.4rem;
     width: 100%;
     background-color: rgba($color: #ffffff, $alpha: 0.3);
-    border-radius: 24px;
+    border-radius: 2.4rem;
     ul {
       display: flex;
       justify-content: center;
-      gap: 30px;
+      gap: 3rem;
       li {
         text-align: center;
         .profile {
           position: relative;
           img {
-            box-shadow: 4px 4px 12px rgb(0, 15, 56, 0.2);
+            box-shadow: 0.4rem 0.4rem 1.2rem rgb(0, 15, 56, 0.2);
             border-radius: 50%;
           }
           em {
             position: absolute;
             bottom: 0;
-            right: -4px;
+            right: -0.4rem;
             display: inline-block;
             text-align: center;
-            width: 22px;
-            height: 22px;
-            line-height: 22px;
+            width: 2.2rem;
+            height: 2.2rem;
+            line-height: 2.2rem;
             border-radius: 50%;
             background-color: #679dff;
             color: #fff;
             font-weight: bold;
-            font-size: 12px;
+            font-size: 1.2rem;
           }
         }
         p {
-          padding-top: 4px;
+          padding-top: 0.4rem;
           font-weight: 700;
-          font-size: 12px;
+          font-size: 1.2rem;
         }
       }
     }
@@ -136,10 +122,10 @@
     display: flex;
     justify-content: space-between;
     width: 100%;
-    height: 580px;
-    padding: 56px;
+    height: 58rem;
+    padding: 5.6rem;
     background-color: #fff;
-    border-radius: 24px;
-    box-shadow: 4px 4px 16px rgba(0, 15, 56, 0.1);
+    border-radius: 2.4rem;
+    box-shadow: 0.4rem 0.4rem 1.6rem rgba(0, 15, 56, 0.1);
   }
 }
